Create README.md
The primary objective of the A/B test was to determine whether the implementation of a banner showcasing key food and drink products on the GloBox mobile website would positively influence user behavior and lead to increased conversions. The control group, comprising users who did not see the banner, was compared to the test group, consisting of users who were presented with the banner upon visiting the website.